+++ packages/playground/.storybook/addons/html/processors/index.ts
@@ -1,2 +1,3 @@
 export * from "./AttributeProcessor";
+export * from "./StylesProcessor";
 export * from "./IProcessor";