+++ packages/playground/_stories/main/ComboBox/ComboBox.stories.ts
@@ -35,6 +35,7 @@ const Template: UI5StoryArgs<ComboBox, StoryArgsSlots> = (
 		args
 ) => html`<ui5-combobox
 		value="${ifDefined(args.value)}"
+		?no-typeahead="${ifDefined(args.noTypeahead)}"
 		placeholder="${ifDefined(args.placeholder)}"
 		?disabled="${ifDefined(args.disabled)}"
 		?readonly="${ifDefined(args.readonly)}"
@@ -94,8 +95,8 @@ Grouping.args = {
 	default: `
 		<ui5-cb-group-item text="A"></ui5-cb-group-item>
 		<ui5-cb-item text="Argentina"></ui5-cb-item>
-		<ui5-cb-item text="Australia"></ui5-cb-item>	
-		<ui5-cb-item text="Austria"></ui5-cb-item>	
+		<ui5-cb-item text="Australia"></ui5-cb-item>
+		<ui5-cb-item text="Austria"></ui5-cb-item>
 		<ui5-cb-group-item text="B"></ui5-cb-group-item>
 		<ui5-cb-item text="Bahrain"></ui5-cb-item>
 		<ui5-cb-item text="Belgium"></ui5-cb-item>
@@ -103,5 +104,15 @@ Grouping.args = {
 		<ui5-cb-group-item text="C"></ui5-cb-group-item>
 		<ui5-cb-item text="Canada"></ui5-cb-item>
 		<ui5-cb-item text="Chile"></ui5-cb-item>
+	`,
+};
+
+export const SuggestionsWrapping = Template.bind({});
+SuggestionsWrapping.args = {
+	placeholder: "Enter product",
+	default: `
+	<ui5-cb-item text="Wireless DSL/ Repeater and Print Server Lorem ipsum dolar st amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor incidunt ut labore et dolore magna aliquyam erat, diam nonumy eirmod tempor individunt ut labore et dolore magna aliquyam erat, sed justo et ea rebum."></ui5-cb-item>
+	<ui5-cb-item text="Widescreen Portable DVD Player w MP3, consetetur sadipscing, sed diam nonumy eirmod tempor invidunt ut labore et dolore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergen, no sea takimata. Tortor pretium viverra suspendisse potenti nullam. Congue quisque egestas diam in arcu cursus.Rutrum tellus pellentesque eu tincidunt tortor. Nec tincidunt praesent semper feugiat nibh sed"></ui5-cb-item>
+	<ui5-cb-item text="Portable DVD Player with 9 inches LCD Monitor"></ui5-cb-item>
 	`,
 };