+++ packages/playground/_stories/main/SplitButton/SplitButton.stories.ts
@@ -38,19 +38,20 @@ const Template: UI5StoryArgs<SplitButton, StoryArgsSlots> = (args) => html`<ui5-
 export const Basic = Template.bind({});
 Basic.args = {
 	default: "Default",
+	accessibleName: "Split Button with Accessible Name",
 };
 
-export const SplitButtonWithMenu = Template.bind(this);
-SplitButtonWithMenu.args = {
+export const OpeningMenu = Template.bind(this);
+OpeningMenu.args = {
 	default: "Open Menu",
 };
-SplitButtonWithMenu.decorators = [
+OpeningMenu.decorators = [
 	(story) => {
 		return html`
 		<ui5-menu id="menuInSplitBtnDefaultAction">
-		<ui5-menu-item text="Edit" icon="add"></ui5-menu-item>
-		<ui5-menu-item text="Save" icon="save"></ui5-menu-item>
-		<ui5-menu-item text="Delete" icon="delete"></ui5-menu-item>
+			<ui5-menu-item text="Edit" icon="add"></ui5-menu-item>
+			<ui5-menu-item text="Save" icon="save"></ui5-menu-item>
+			<ui5-menu-item text="Delete" icon="delete"></ui5-menu-item>
 		</ui5-menu>
 	${story()}
 	<script>
@@ -62,28 +63,10 @@ SplitButtonWithMenu.decorators = [
 	</script>`;}
 ];
 
-export const Disabled = Template.bind({});
-Disabled.storyName = "Disabled SplitButton";
-Disabled.args = {
-	default: "Disabled",
-	disabled: true,
-};
-
-export const Design = Template.bind({});
-Design.args = {
-	default: "Attention",
-	design: ButtonDesign.Attention,
-};
-
-export const WithIcon = Template.bind({});
-WithIcon.args = {
-	default: "Icon",
-	icon: "add",
-};
-
-export const WithActiveIcon = Template.bind({});
-WithActiveIcon.args = {
-	default: "Press Me",
-	icon: "add",
-	activeIcon: "accept",
-};
+export const DifferentDesigns: StoryFn = () => html`
+	<ui5-split-button design="${ButtonDesign.Emphasized}"> Emphasized </ui5-split-button>
+	<ui5-split-button design="${ButtonDesign.Attention}"> Attention </ui5-split-button>
+	<ui5-split-button design="${ButtonDesign.Positive}"> Positive </ui5-split-button>
+	<ui5-split-button design="${ButtonDesign.Negative}"> Negative </ui5-split-button>
+	<ui5-split-button design="${ButtonDesign.Transparent}"> Transparent </ui5-split-button>
+`;